Add GPUArrays extension
