+++ DifferentiationInterface/ext/DifferentiationInterfaceGTPSAExt/onearg.jl
@@ -46,7 +46,7 @@ function DI.pushforward(
     DI.check_prep(f, prep, backend, x, tx, contexts...)
     fc = DI.fix_tail(f, map(DI.unwrap, contexts)...)
     ty = map(tx) do dx
-        foreach((t, xi, dxi) -> (t[0] = xi; t[1] = dxi), prep.xt, x, dx)
+        foreach((t, xi, dxi) -> (t[0]=xi; t[1]=dxi), prep.xt, x, dx)
         yt = fc(prep.xt)
         if yt isa Number
             return yt[1]
@@ -71,7 +71,7 @@ function DI.pushforward!(
     fc = DI.fix_tail(f, map(DI.unwrap, contexts)...)
     for b in eachindex(tx, ty)
         dx, dy = tx[b], ty[b]
-        foreach((t, xi, dxi) -> (t[0] = xi; t[1] = dxi), prep.xt, x, dx)
+        foreach((t, xi, dxi) -> (t[0]=xi; t[1]=dxi), prep.xt, x, dx)
         yt = fc(prep.xt)
         map!(t -> t[1], dy, yt)
     end
